+++ CHANGES.rst
@@ -10,6 +10,8 @@ Changelog
 * [snap]: Add support for ``snap`` on Linux.
 * [mpm]: Add ``--time`` / ``--no-time`` flag to show elapsed execution time.
   Closes #9.
+* [mpm]: Print table rendering, stats and timing in console output instead of
+  logger to allow them to be greppable.
 * [mpm]: Replace internal helpers with upstreamed ``boltons`` utils.
 * [mpm]: Allow for manager-specific search path to help hunting down CLIs.
 * [mpm]: Bump requirement to ``boltons 20.1.1``.